[DOC-214] Migrate Alfresco to AWS S3 (#9)
* [DOC-214] Sending attachment through RabbitMQ to doc-file-service

* [DOC-214] Add get file for previewing, outgoing document now can save attachment

* [DOC-214] Migrate Alfresco to AWS S3

+++ service/src/main/java/edu/hcmus/doc/fileservice/model/dto/FileDto.java
@@ -15,7 +15,7 @@ public class FileDto {
   private String modifiedDate;
   private String creator;
   private String modifier;
-  private Double size;
+  private Long size;
   private String mimeType;
   private String url;
   private byte[] data;
